🏷️rename: 상수 파일 이름 변경 및 적용
