+++ src/app/user/[id]/setting/edit/components/EditNickname/EditNickname.tsx
@@ -7,8 +7,9 @@ import { Button, Input, InputLabel } from "@/components";
 import useTheme from "@/lib/hooks/useTheme";
 import { FONT_SIZE, FONT_WEIGHT } from "@/styles/theme";
 import { USER_EDIT_PLACEHOLDER } from "../../UserEdit.constants";
-import { validation_user } from "@/lib/constants/userValidate";
+
 import { MouseEvent } from "react";
+import { validation_user } from "@/lib/constants/userConstants";
 
 interface EditNicknameProps {
   register: UseFormRegister<UserEditData>;
